Refactor template package
refactor template package to have all the templating functions
in a different package named `templatefuncs` and template
package should have only those exposed functions which in turn
make use of these templating functions.
