+++ pkg/templatefuncs/v1alpha1/runtask_functions.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package template
+package templatefuncs
 
 import (
 	"encoding/json"
